Renamed IDocumentStoring into IDocumentService. Moved DocumentManager.Conracts interfaces into /interfaces Folder
