--- src/Data/LuceneRepository/Logic/IQueryController.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Mame.Doci.Data.LuceneRepository.Logic
-{
-    interface IQueryController
-    {
-
-
-    }
-}
