--- src/Data/LuceneRepository/Logic/QueryController.cs
@@ -1,15 +1,0 @@
-﻿using Mame.Doci.Logic.DocumentManager.Contracts;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Mame.Doci.Data.LuceneRepository.Logic
-{
-    class QueryController: IQueryController, IDocumentRequesting
-    {
-
-
-
-    }
-}
