+++ src/Data/LuceneRepository/LuceneRepository.csproj
@@ -189,8 +189,6 @@
     <Compile Include="Data\IndexImportFile.cs" />
     <Compile Include="Logic\IIndexingController.cs" />
     <Compile Include="Logic\IndexingController.cs" />
-    <Compile Include="Logic\IQueryController.cs" />
-    <Compile Include="Logic\QueryController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
