--- src/Logic/DocumentManager.Contracts/IDocumentRequesting.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace Mame.Doci.Logic.DocumentManager.Contracts
-{
-    public interface IDocumentRequesting
-    {
-    }
-}
