+++ src/Logic/DocumentManager.Contracts/Interfaces/IDocumentRepository.cs
@@ -4,11 +4,11 @@ using System.IO;
 using System.Linq;
 using System.Text;
 
-namespace Mame.Doci.Logic.DocumentManager.Contracts
+namespace Mame.Doci.Logic.DocumentManager.Contracts.Interfaces
 {
 
 
-    public interface IDocumentStoring
+    public interface IDocumentRepository
     {
         void Store (FileInfo storeFile);        // throws DocumentStoreException
         void Store (List<FileInfo> storeFiles);// throws DocumentStoreException
