+++ src/Logic/DocumentManager.Tests/UnitTests/Storing/StoringControllerTests.LogMessage.cs
@@ -5,7 +5,7 @@ using NSubstitute;
 using Mame.Doci.CrossCutting.Logging.Contracts;
 using Mame.Doci.Logic.DocumentManager.Storing;
 using NUnit.Framework;
-using Mame.Doci.Logic.DocumentManager.Contracts;
+using Mame.Doci.Logic.DocumentManager.Contracts.Interfaces;
 
 namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
 {
@@ -16,7 +16,7 @@ namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
         {
             //ARRANGE
             List<FileInfo> fileinfos = new List<FileInfo> ();
-            IDocumentStoring docStorer = null;
+            IDocumentRepository docStorer = null;
             var theController = new StoringController (docStorer);
 
             //ACT
@@ -31,7 +31,7 @@ namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
         {
             //ARRANGE
             List<FileInfo> fileinfos = new List<FileInfo> ();
-            IDocumentStoring docStorer = null;
+            IDocumentRepository docStorer = null;
             var theController = new StoringController (docStorer);
 
             var theLogger = Substitute.For<ILogger> ();
