+++ src/Logic/DocumentManager.Tests/UnitTests/Storing/StoringControllerTests.UserWantsToStore.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using Mame.Doci.Logic.DocumentManager.Contracts;
+using Mame.Doci.Logic.DocumentManager.Contracts.Interfaces;
 using Mame.Doci.Logic.DocumentManager.Storing;
 using NUnit.Framework;
 
@@ -15,7 +15,7 @@ namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
         {
             //ARRANGE
             List<FileInfo> fileinfos = new List<FileInfo> ();
-            IDocumentStoring docStorer = null;
+            IDocumentRepository docStorer = null;
             var theController = new StoringController (docStorer);
                        
 
@@ -31,7 +31,7 @@ namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
         {
             //ARRANGE
             FileInfo fileinfo = new FileInfo (@"C:\thisFileIsNotExisting.txt");
-            IDocumentStoring docStorer = null;
+            IDocumentRepository docStorer = null;
             var theController = new StoringController (docStorer);
 
             //ACT
@@ -46,7 +46,7 @@ namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
         {
             //ARRANGE
             FileInfo nullFileInfo = null;
-            IDocumentStoring docStorer = NSubstitute.Substitute.For<IDocumentStoring> ();
+            IDocumentRepository docStorer = NSubstitute.Substitute.For<IDocumentRepository> ();
             var theController = new StoringController (docStorer);
             //ACT
             //ASSERT
@@ -60,7 +60,7 @@ namespace Mame.Doci.Logic.DocumentAccessing.Tests.UnitTests.Storing
         {
             //ARRANGE
             List<FileInfo> nullFileInfos = null;
-            IDocumentStoring docStorer = NSubstitute.Substitute.For<IDocumentStoring> ();
+            IDocumentRepository docStorer = NSubstitute.Substitute.For<IDocumentRepository> ();
             var theController = new StoringController (docStorer);
 
             //ACT
