+++ src/Logic/DocumentManager/Factories/StoringControllerFactory.cs
@@ -1,11 +1,11 @@
 ﻿using Mame.Doci.CrossCutting.Logging.Contracts;
-using Mame.Doci.Logic.DocumentManager.Contracts;
+using Mame.Doci.Logic.DocumentManager.Contracts.Interfaces;
 
 namespace Mame.Doci.Logic.DocumentManager.Storing.Factories
 {
     public class StoringControllerFactory
     {
-        public static IStoringForUser CreateDefault (IDocumentStoring luceneStorer, ILogger logger = null)
+        public static IDocumentService CreateDefault (IDocumentRepository luceneStorer, ILogger logger = null)
         {
             //IDocumentStoring luceneStorer = LuceneIndexingControllerFactory.CreateDefault (logger);
             return new StoringController (luceneStorer) {Logger=logger};
